Implementation of the CultureOrDefaultConvert.ConvertBack method
I think that CultureOrDefaultConverter needs to implement ConvertBack method in order to swtich the value of Translator.Culture in ComboBox like "1.1. Basic usage".
If ConvertBack method is not implemented, the TwoWay binding will not work and "1.1. Basic usage" example will not work with runtime errors.
I thought to simply return value parameter, but I felt that the same implementation as Convert is optimal.
This will work "1.1. Basic usage" example.