UPDATE README: Modify the description of the change of Culture.
README are very important for users, so I would like to fix incorrect descriptions.

Since Translator.CurrentCulture is private, it can not be assigned, resulting in a compile error.
Change to Translator.Culture which can be assigned.

It is the reflection of the rename (ea01b544cbd61b1368c8e22c0df14696ecb5e7db Translator.CurrentCulture -> Translator.Culture).